feat: packed attestation (#64)

+++ apps/nestjs/src/controllers/WebAuthnCredentials.controller.ts
@@ -1,10 +1,23 @@
+import { KeyClient } from '@azure/keyvault-keys';
 import { Controller, UseFilters, UseGuards } from '@nestjs/common';
-import { contract } from '@repo/contract';
-import { KeyVault } from '@repo/key-vault';
+import { AuditLog } from '@repo/audit-log';
+import { AuditLogAction, AuditLogEntity } from '@repo/audit-log/enums';
+import { Permission, TokenType } from '@repo/auth/enums';
+import type { JwtPayload } from '@repo/auth/validation';
+import { nestjsContract } from '@repo/contract/nestjs';
+import {
+  DeleteWebAuthnCredentialResponseSchema,
+  GetWebAuthnCredentialResponseSchema,
+  ListWebAuthnCredentialsResponseSchema,
+} from '@repo/contract/validation';
+import { Forbidden } from '@repo/exception/http';
 import { Logger } from '@repo/logger';
+import { Pagination } from '@repo/pagination';
 import { WebAuthnCredentialKeyMetaType } from '@repo/prisma';
-import { WebAuthnCredential, type JwtPayload } from '@repo/validation';
+import { WebAuthnCredentialWithMeta } from '@repo/virtual-authenticator/types';
+import { WebAuthnCredential } from '@repo/virtual-authenticator/validation';
 import { tsRestHandler, TsRestHandler } from '@ts-rest/nest';
+import { Schema } from 'effect';
 
 import { Jwt } from '../decorators/Jwt.decorator';
 import { ExceptionFilter } from '../filters/Exception.filter';
@@ -16,51 +29,72 @@ import { PrismaService } from '../services/Prisma.service';
 export class WebAuthnCredentialsController {
   constructor(
     private readonly prisma: PrismaService,
-    private readonly keyVault: KeyVault,
+    private readonly keyClient: KeyClient,
     private readonly logger: Logger,
+    private readonly auditLog: AuditLog,
   ) {}
 
-  @TsRestHandler(contract.api.webAuthnCredentials.list)
+  @TsRestHandler(nestjsContract.api.webAuthnCredentials.list)
   @UseGuards(AuthenticatedGuard)
   async listWebAuthnCredentials(@Jwt() jwtPayload: JwtPayload) {
-    return tsRestHandler(contract.api.webAuthnCredentials.list, async () => {
-      const { user } = jwtPayload;
+    return tsRestHandler(
+      nestjsContract.api.webAuthnCredentials.list,
+      async ({ query }) => {
+        const { userId, permissions } = jwtPayload;
 
-      const webAuthnCredentials = await this.prisma.webAuthnCredential.findMany(
-        {
-          where: {
-            userId: user.id,
-            webAuthnCredentialKeyMetaType:
-              WebAuthnCredentialKeyMetaType.KEY_VAULT,
-          },
-          include: {
-            webAuthnCredentialKeyVaultKeyMeta: true,
-          },
-        },
-      );
+        if (!permissions.includes(Permission['WebAuthnCredential.read'])) {
+          throw new Forbidden();
+        }
 
-      return {
-        status: 200,
-        body: contract.api.webAuthnCredentials.list.responses[200].encode(
-          webAuthnCredentials as WebAuthnCredential[],
-        ),
-      };
-    });
+        const pagination = new Pagination(async ({ pagination }) => {
+          const webAuthnCredentials =
+            await this.prisma.webAuthnCredential.findMany({
+              where: {
+                userId: userId,
+                webAuthnCredentialKeyMetaType:
+                  WebAuthnCredentialKeyMetaType.KEY_VAULT,
+              },
+              include: {
+                webAuthnCredentialKeyVaultKeyMeta: true,
+              },
+              ...pagination,
+            });
+
+          return webAuthnCredentials as WebAuthnCredentialWithMeta[];
+        });
+
+        const result = await pagination.fetch({
+          limit: query.limit,
+          cursor: query.cursor,
+        });
+
+        return {
+          status: 200,
+          body: Schema.encodeUnknownSync(ListWebAuthnCredentialsResponseSchema)(
+            result,
+          ),
+        };
+      },
+    );
   }
 
-  @TsRestHandler(contract.api.webAuthnCredentials.get)
+  @TsRestHandler(nestjsContract.api.webAuthnCredentials.get)
   @UseGuards(AuthenticatedGuard)
   async getWebAuthnCredential(@Jwt() jwtPayload: JwtPayload) {
     return tsRestHandler(
-      contract.api.webAuthnCredentials.get,
+      nestjsContract.api.webAuthnCredentials.get,
       async ({ params }) => {
-        const { user } = jwtPayload;
+        const { userId, permissions } = jwtPayload;
+
+        if (!permissions.includes(Permission['WebAuthnCredential.read'])) {
+          throw new Forbidden();
+        }
 
         const webAuthnCredential =
           await this.prisma.webAuthnCredential.findUniqueOrThrow({
             where: {
               id: params.id,
-              userId: user.id,
+              userId: userId,
             },
             include: {
               webAuthnCredentialKeyVaultKeyMeta: true,
@@ -69,7 +103,7 @@ export class WebAuthnCredentialsController {
 
         return {
           status: 200,
-          body: contract.api.webAuthnCredentials.get.responses[200].encode(
+          body: Schema.encodeSync(GetWebAuthnCredentialResponseSchema)(
             webAuthnCredential as WebAuthnCredential,
           ),
         };
@@ -77,18 +111,22 @@ export class WebAuthnCredentialsController {
     );
   }
 
-  @TsRestHandler(contract.api.webAuthnCredentials.delete)
+  @TsRestHandler(nestjsContract.api.webAuthnCredentials.delete)
   @UseGuards(AuthenticatedGuard)
   async deleteWebAuthnCredential(@Jwt() jwtPayload: JwtPayload) {
     return tsRestHandler(
-      contract.api.webAuthnCredentials.delete,
+      nestjsContract.api.webAuthnCredentials.delete,
       async ({ params }) => {
-        const { user } = jwtPayload;
+        const { userId, permissions } = jwtPayload;
+
+        if (!permissions.includes(Permission['WebAuthnCredential.delete'])) {
+          throw new Forbidden();
+        }
 
         const webAuthnCredential = await this.prisma.webAuthnCredential.delete({
           where: {
             id: params.id,
-            userId: user.id,
+            userId: userId,
           },
           include: {
             webAuthnCredentialKeyVaultKeyMeta: true,
@@ -97,25 +135,37 @@ export class WebAuthnCredentialsController {
 
         this.logger.debug('Removing WebAuthnCredential.', {
           webAuthnCredential,
-          userId: jwtPayload.id,
+          userId: userId,
         });
 
         if (
           webAuthnCredential.webAuthnCredentialKeyMetaType ===
           WebAuthnCredentialKeyMetaType.KEY_VAULT
         ) {
-          await this.keyVault.deleteKey({
-            keyName:
-              webAuthnCredential.webAuthnCredentialKeyVaultKeyMeta!
-                .keyVaultKeyName,
-          });
+          const pollOperation = await this.keyClient.beginDeleteKey(
+            webAuthnCredential.webAuthnCredentialKeyVaultKeyMeta!
+              .keyVaultKeyName,
+          );
+
+          await pollOperation.pollUntilDone();
         }
+
+        await this.auditLog.audit({
+          action: AuditLogAction.DELETE,
+          entity: AuditLogEntity.WEBAUTHN_CREDENTIAL,
+
+          apiKeyId:
+            jwtPayload.tokenType === TokenType.API_KEY
+              ? jwtPayload.apiKeyId
+              : undefined,
+          userId: jwtPayload.userId,
+        });
 
         return {
           status: 200,
-          body: contract.api.webAuthnCredentials.delete.responses[200].encode({
-            success: true,
-          }),
+          body: Schema.encodeSync(DeleteWebAuthnCredentialResponseSchema)(
+            webAuthnCredential as WebAuthnCredential,
+          ),
         };
       },
     );
